Added Feature Level 9 support

+++ src/libGLESv2/libGLESv2.cpp
@@ -16,6 +16,7 @@
 #include "libGLESv2/Fence.h"
 #include "libGLESv2/Framebuffer.h"
 #include "libGLESv2/Renderbuffer.h"
+#include "libGLESv2/Renderer/Renderer.h"
 #include "libGLESv2/Program.h"
 #include "libGLESv2/ProgramBinary.h"
 #include "libGLESv2/Texture.h"
@@ -185,7 +186,14 @@ void __stdcall glBindBuffer(GLenum target, GLuint buffer)
             context->bindGenericUniformBuffer(buffer);
             return;
           case GL_TRANSFORM_FEEDBACK_BUFFER:
-            context->bindGenericTransformFeedbackBuffer(buffer);
+              if (context->getRenderer()->getMaxTransformFeedbackBuffers() > 0)
+              {
+                  context->bindGenericTransformFeedbackBuffer(buffer);
+              }
+              else
+              {
+                  return gl::error(GL_INVALID_OPERATION);
+              }
             return;
           default:
             return gl::error(GL_INVALID_ENUM);
@@ -2382,6 +2390,17 @@ void __stdcall glGetProgramiv(GLuint program, GLenum pname, GLint* params)
               case GL_TRANSFORM_FEEDBACK_VARYINGS:
               case GL_TRANSFORM_FEEDBACK_VARYING_MAX_LENGTH:
                 return gl::error(GL_INVALID_ENUM);
+            }
+        }
+
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
+        {
+            switch (pname)
+            {
+            case GL_TRANSFORM_FEEDBACK_BUFFER_MODE:
+            case GL_TRANSFORM_FEEDBACK_VARYINGS:
+            case GL_TRANSFORM_FEEDBACK_VARYING_MAX_LENGTH:
+                return gl::error(GL_INVALID_OPERATION);
             }
         }
 
@@ -5661,6 +5680,17 @@ void __stdcall glGetIntegeri_v(GLenum target, GLuint index, GLint* data)
             return gl::error(GL_INVALID_OPERATION);
         }
 
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
+        {
+            switch (target)
+            {
+              case GL_TRANSFORM_FEEDBACK_BUFFER_START:
+              case GL_TRANSFORM_FEEDBACK_BUFFER_SIZE:
+              case GL_TRANSFORM_FEEDBACK_BUFFER_BINDING:
+                  return gl::error(GL_INVALID_OPERATION);
+            }
+        }
+
         switch (target)
         {
           case GL_TRANSFORM_FEEDBACK_BUFFER_START:
@@ -5726,6 +5756,11 @@ void __stdcall glBeginTransformFeedback(GLenum primitiveMode)
             return gl::error(GL_INVALID_OPERATION);
         }
 
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
+        {
+            return gl::error(GL_INVALID_OPERATION);
+        }
+
         switch (primitiveMode)
         {
           case GL_TRIANGLES:
@@ -5764,6 +5799,11 @@ void __stdcall glEndTransformFeedback(void)
     if (context)
     {
         if (context->getClientVersion() < 3)
+        {
+            return gl::error(GL_INVALID_OPERATION);
+        }
+
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
         {
             return gl::error(GL_INVALID_OPERATION);
         }
@@ -5865,6 +5905,14 @@ void __stdcall glBindBufferBase(GLenum target, GLuint index, GLuint buffer)
             return gl::error(GL_INVALID_OPERATION);
         }
 
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
+        {
+            if (target == GL_TRANSFORM_FEEDBACK_BUFFER)
+            {
+                return gl::error(GL_INVALID_OPERATION);
+            }
+        }
+
         switch (target)
         {
           case GL_TRANSFORM_FEEDBACK_BUFFER:
@@ -5913,6 +5961,11 @@ void __stdcall glTransformFeedbackVaryings(GLuint program, GLsizei count, const 
     if (context)
     {
         if (context->getClientVersion() < 3)
+        {
+            return gl::error(GL_INVALID_OPERATION);
+        }
+
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
         {
             return gl::error(GL_INVALID_OPERATION);
         }
@@ -7178,6 +7231,17 @@ void __stdcall glGetInteger64i_v(GLenum target, GLuint index, GLint64* data)
             return gl::error(GL_INVALID_OPERATION);
         }
 
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
+        {
+            switch (target)
+            {
+              case GL_TRANSFORM_FEEDBACK_BUFFER_START:
+              case GL_TRANSFORM_FEEDBACK_BUFFER_SIZE:
+              case GL_TRANSFORM_FEEDBACK_BUFFER_BINDING:
+                return gl::error(GL_INVALID_OPERATION);
+            }
+        }
+
         switch (target)
         {
           case GL_TRANSFORM_FEEDBACK_BUFFER_START:
@@ -7543,6 +7607,11 @@ void __stdcall glBindTransformFeedback(GLenum target, GLuint id)
             return gl::error(GL_INVALID_OPERATION);
         }
 
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
+        {
+            return gl::error(GL_INVALID_OPERATION);
+        }
+
         switch (target)
         {
           case GL_TRANSFORM_FEEDBACK:
@@ -7583,6 +7652,11 @@ void __stdcall glDeleteTransformFeedbacks(GLsizei n, const GLuint* ids)
             return gl::error(GL_INVALID_OPERATION);
         }
 
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
+        {
+            return gl::error(GL_INVALID_OPERATION);
+        }
+
         for (int i = 0; i < n; i++)
         {
             context->deleteTransformFeedback(ids[i]);
@@ -7599,6 +7673,11 @@ void __stdcall glGenTransformFeedbacks(GLsizei n, GLuint* ids)
     if (context)
     {
         if (context->getClientVersion() < 3)
+        {
+            return gl::error(GL_INVALID_OPERATION);
+        }
+
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
         {
             return gl::error(GL_INVALID_OPERATION);
         }
@@ -7642,6 +7721,11 @@ void __stdcall glPauseTransformFeedback(void)
             return gl::error(GL_INVALID_OPERATION);
         }
 
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
+        {
+            return gl::error(GL_INVALID_OPERATION);
+        }
+
         gl::TransformFeedback *transformFeedback = context->getState().getCurrentTransformFeedback();
         ASSERT(transformFeedback != NULL);
 
@@ -7664,6 +7748,11 @@ void __stdcall glResumeTransformFeedback(void)
     if (context)
     {
         if (context->getClientVersion() < 3)
+        {
+            return gl::error(GL_INVALID_OPERATION);
+        }
+
+        if (context->getRenderer()->getMaxTransformFeedbackBuffers() == 0)
         {
             return gl::error(GL_INVALID_OPERATION);
         }
